backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
pydantic-settings==2.2.1
python-dotenv==1.0.1
alembic==1.13.1
celery==5.3.6
redis==5.0.1
python-multipart==0.0.9
pytest==8.0.2
httpx==0.27.0
PyExifTool==0.5.6
Pillow==10.2.0
pgvector==0.2.5
opentelemetry-api==1.23.0
opentelemetry-sdk==1.23.0
opentelemetry-instrumentation-fastapi==0.44b0
opentelemetry-instrumentation-celery==0.44b0
opentelemetry-exporter-otlp-proto-http==1.23.0
opentelemetry-exporter-otlp-proto-grpc==1.23.0
opentelemetry-instrumentation-sqlalchemy==0.44b0
opentelemetry-instrumentation-httpx==0.44b0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
